+++ ProjectManager/ProjectManager/View/TaskListViewModule/TaskListContainerView.swift
@@ -17,47 +17,12 @@ struct TaskListContainerView: View {
     var body: some View {
         VStack(alignment: .leading) {
             TaskListTitleView(taskType: taskType)
-            taskList
+            TaskListView(taskType: taskType, isShowSheet: $isShowSheet)
         }
-    }
-    
-    private var taskList: some View {
-        List {
-            ForEach(tasks) { task in
-                Button(action: {
-                    self.isShowSheet.toggle()
-                }) {
-                    TaskRowView(task: task)
-                        .contextMenu {
-                            contextMenuView(task, taskType)
-                        }
-                        .sheet(isPresented: $isShowSheet, onDismiss: nil) {
-                            TaskDetailView(
-                                task: task,
-                                isShowSheet: $isShowSheet
-                            )
-                        }
-                }
-            }
-            .onDelete { indexSet in
-                viewModel.remove(tasks[indexSet])
-            }
-        }
-        
     }
     
     private var tasks: [Task] {
         viewModel.findTasks(of: taskType)
-    }
-    
-    private func contextMenuView(_ task: Task, _ taskType: TaskStatus) -> some View {
-        ForEach(TaskStatus.allCases) { status in
-            if status != taskType {
-                Button(action: { viewModel.update(task, taskStatus: status) }) {
-                    Text("Move to \(status.description)")
-                }
-            }
-        }
     }
     
 }
